+++ components/FilterInputComponent/FilterInputToken.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react";
+
+export const FilterInputToken: React.FC<PropsWithChildren> = ({ children }) => (
+  <span className="animate-in fade-in pl-2 font-bold">{children}</span>
+);
